wip(markdown): wave 2 — renderer.ts verbatim migrated to @fleet-console/markdown/core
- Extract renderMarkdown/encodeMermaidSource/decodeMermaidSource/TocItem to markdown/core.ts
- Add resolveWikiLink DI option; wiki patterns fall through as escape text when not injected
- inline escapeHtml as module-private (codex/utils/html.ts unchanged per constraint 1)
- Update reading-controller to inject resolveWikiLink: (id) => entryPath(id)
- Update toc-sheet.ts TocItem import path to @fleet-console/markdown/core
- Update diagrams.ts decodeMermaidSource import to @fleet-console/markdown/core
- Delete codex/markdown/renderer.ts
- Update security tests to use @fleet-console/markdown/core
- Add @fleet-console/markdown workspace:* to console + file-explorer package.json
